+++ src/backend/demo/tests/test_commands_create_demo.py
@@ -26,8 +26,11 @@ def test_commands_create_demo():
     """The create_demo management command should create objects as expected."""
     call_command("create_demo")
 
-    assert models.Service.objects.exclude(name="docs").count() == 2
+    assert models.Service.objects.exclude(name="docs").count() == 3
     assert opensearch_client().count()["count"] == 4
 
     docs = models.Service.objects.get(name="docs")
     assert docs.client_id == "impress"
+
+    drive = models.Service.objects.get(name="drive")
+    assert drive.client_id == "drive"