+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to
 - ✨(backend) limit search to the calling app (audience) and a configured
               list of services
 - 🔧(compose) rename docker network 'lasuite-net' as 'lasuite-network'
+- ✨(backend) add demo service for Drive.